Loading data from Mercuto is now delayed until first process
